requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
environs==9.5.0
marshmallow==3.15.0
packaging==21.3
pyparsing==3.0.7
python-dotenv==0.19.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
google-cloud-dialogflow==2.13.0
